Refactor providers + yubihsm-rs update + Sync-safe signers
* Factors providers to toplevel of project, since when we go
  multi-algorithm one provider may implement severl algorithms.
  Frees up the `ed25519/` directory so we can split up `mod.rs`
* Updates to yubihsm-rs 0.6.0 which provides a thread-safe API.
  Splits up YubiHSM sessions and signers so several signers can
  share a single session.
* Since yubihsm-rs is now thread-safe, bound ed25519::Signer on
  Sync. Signatory now provides a thread-safe API.
