+++ Cargo.toml
@@ -19,13 +19,15 @@ ed25519-dalek = { version = "= 0.6.0", optional = true, default-features = false
 failure = { version = "0.1", default-features = false }
 failure_derive = { version = "0.1", default-features = false }
 sha2 = { version = "0.6", optional = true }
-yubihsm = { version = "0.1", optional = true }
+yubihsm = { version = "0.6", optional = true }
 
 [features]
 dalek-provider = ["curve25519-dalek", "ed25519-dalek", "sha2"]
 default = ["dalek-provider"]
-yubihsm-provider = ["yubihsm"]
+std = ["failure/std", "failure_derive/std"]
+yubihsm-provider = ["std", "yubihsm"]
 yubihsm-mockhsm = ["yubihsm-provider", "yubihsm/mockhsm"]
 
 [package.metadata.docs.rs]
-rustc-args = ["-C", "target-feature=+aes"]
+features = ["dalek-provider", "yubihsm-provider"]
+rustc-args = ["-Ctarget-feature=+aes"]
